undo dmer buttons and add access level permission
