+++ medical-portal/src/UI/src/app/dashboard/dashboard.component.html
@@ -157,8 +157,64 @@
                     </div>
                   </div>
                   <div>
-                    <app-dmer-buttons (afterDialogClosed)="updateDmerCases()" [searchedCase]="searchedCase"></app-dmer-buttons>
-                  </div>
+                    <div class="search-buttons">
+                      @if (
+                        accessLevel === Role.Practitioner &&
+                        searchedCase.dmerStatus === DMERStatusEnum.RequiredUnclaimed
+                      ) {
+                        <button
+                          mat-flat-button
+                          class="primaryButton"
+                          (click)="openClaimPopup(searchedCase)"
+                        >
+                          Claim Dmer
+                        </button>
+                      }
+
+                      @if (
+                        searchedCase.dmerStatus ===
+                        DMERStatusEnum.RequiredClaimed
+                      ) {
+                        <button
+                          mat-flat-button
+                          class="primaryButton"
+                          (click)="openClaimPopup(searchedCase)"
+                        >
+                          Unclaim Dmer
+                        </button>
+                      }
+
+                      @if (
+                        searchedCase.dmerStatus ===
+                          DMERStatusEnum.RequiredUnclaimed &&
+                        searchedCase.isOwner == true
+                      ) {
+                        <button
+                          mat-flat-button
+                          class="primaryButton"
+                          (click)="openPopup()"
+                        >
+                          Start DMER
+                        </button>
+                      }
+
+                      @if (
+                        searchedCase.dmerStatus ===
+                          DMERStatusEnum.RequiredClaimed ||
+                        searchedCase.dmerStatus ===
+                          DMERStatusEnum.NonComplyClaimed
+                      ) {
+                        <div>
+                          <button
+                            mat-flat-button
+                            class="primaryButton"
+                            (click)="openPopup()"
+                          >
+                            Continue DMER
+                          </button>
+                        </div>
+                      }
+                    </div>                  </div>
                 </div>
               </div>
               <div style="padding-left: 18px">
